recipes-phosphor: Enable SOL host console
This patch enables Serial over LAN Host console.
It uses "ttyS0" as a terminal device and sets the baud rate to 115200.

SOL is not verified on Onyx as we don't have a SoC or a mechanism to generate
UART traffic. Will try to verify with IOD system.

Note that since all the SP5 platforms use Hawaii based design and from BMC
side its the same serial device, keeping a single server config file for all
platforms. If required this can be amended per platform basis in future.

Signed-off-by: Sreedhar Govardhana <srgovard@amd.com>
Change-Id: Id9ab371c92a5cc1566df3fed1833bc6f2a87fe93